+++ include/HermiteCurve.h
@@ -5,12 +5,12 @@
 
 class HermiteCurve {
 	public:
-		Vec3f ps;
-		Vec3f pe;
-		Vec3f ts;
-		Vec3f te;
+		Vec3d ps;
+		Vec3d pe;
+		Vec3d ts;
+		Vec3d te;
 
-		HermiteCurve(Vec3f _ps, Vec3f _pe, Vec3f _ts, Vec3f _te){
+		HermiteCurve(Vec3d _ps, Vec3d _pe, Vec3d _ts, Vec3d _te){
 			ps = _ps;
 			pe = _pe;
 			ts = _ts;
@@ -18,20 +18,20 @@ class HermiteCurve {
 		}
 
 		// s in [0;1]
-		Vec3f interpolate(float s){
-			Vec3f _value = (2.f*s*s*s - 3.f*s*s +1.f)*ps +
-				(s*s*s - 2.f*s*s +s)*ts +
-				(-2.f*s*s*s + 3.f*s*s)*pe +
+		Vec3d interpolate(double s){
+			Vec3d _value = (2.0*s*s*s - 3.0*s*s +1.0)*ps +
+				(s*s*s - 2.0*s*s +s)*ts +
+				(-2.0*s*s*s + 3.0*s*s)*pe +
 				(s*s*s -s*s)*te;
 
 			return _value;
 		}
 
-		Vec3f get_tangent(float s){
-			Vec3f _value = (6.f*s*s - 6.f*s)*ps +
-				(3*s*s - 4.f*s +1.f)*ts +
-				(-6.f*s*s + 6.f*s)*pe +
-				(3.f*s*s -2.f*s)*te;
+		Vec3d get_tangent(double s){
+			Vec3d _value = (6.0*s*s - 6.0*s)*ps +
+				(3.0*s*s - 4.0*s +1.0)*ts +
+				(-6.0*s*s + 6.0*s)*pe +
+				(3.0*s*s -2.0*s)*te;
 
 			return _value;
 		}
